requirements.txt
# Tested with Python 3.12.1 and pip 23.2.1

requests
pytz
svg.py
python-dotenv
